pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-secops"
version = "0.1.0"
description = "Python SDK for Google SecOps"
readme = "README.md"
requires-python = ">=3.7"
license = "Apache-2.0"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-api-python-client>=2.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "tox>=3.24.0",
]
docs = [
    "sphinx>=4.0.0",
    "sphinx-rtd-theme>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=google_secops"
markers = [
    "integration: marks tests as integration tests that interact with real APIs",
] 